refactoring the main screen
