+++ src/main/kotlin/org/bizilabs/apps/shaker/components/CursorField.kt
@@ -1,0 +1,46 @@
+package org.bizilabs.apps.shaker.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Text
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.PointerEventType
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun CursorField(modifier: Modifier = Modifier, onScroll: (Float) -> Unit) {
+    Card(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp)
+            .padding(bottom = 16.dp),
+        colors = CardDefaults.cardColors(),
+        shape = RoundedCornerShape(10),
+    ) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .pointerInput(Unit) {
+                    awaitPointerEventScope {
+                        while (true) {
+                            val event = awaitPointerEvent()
+                            if (event.type == PointerEventType.Scroll) {
+                                val deltaY = event.changes.first().scrollDelta.y
+                                onScroll(deltaY)
+                            }
+                        }
+                    }
+                },
+        ) {
+            Text(text = "", modifier = Modifier.align(Alignment.Center))
+        }
+    }
+}
